Refactor BubbleView environment initialization
