+++ GliaWidgets/Sources/Component/Bubble/BubbleView.swift
@@ -167,12 +167,3 @@ class BubbleView: BaseView {
         gesture.setTranslation(.zero, in: self)
     }
 }
-
-extension BubbleView {
-    struct Environment {
-        var data: FoundationBased.Data
-        var uuid: () -> UUID
-        var gcd: GCD
-        var imageViewCache: ImageView.Cache
-    }
-}
